testing/mochitest/tests/Harness_sanity/test_SpecialPowersSpawn.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for SpecialPowers.spawn</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<iframe id="iframe"></iframe>

<span id="hello">World.</span>

<script>
/* eslint-disable prettier/prettier */
/* globals SpecialPowers, content */

  add_task(async function() {
    let frame = document.getElementById("iframe");
    frame.src = "https://example.com/tests/testing/mochitest/tests/Harness_sanity/file_spawn.html";

    await new Promise(resolve => {
      frame.addEventListener("load", resolve, {once: true});
    });

    let result = await SpecialPowers.spawn(frame, ["#span"], selector => {
      let elem = content.document.querySelector(selector);
      return elem.textContent;
    });

    is(result, "Hello there.", "Got correct element text from frame");

    result = await SpecialPowers.spawn(frame, ["#hello"], selector => {
      return SpecialPowers.spawn(content.parent, [selector], selector => {
        let elem = content.document.querySelector(selector);
        return elem.textContent;
      });
    });

    is(result, "World.", "Got correct element text from frame's window.parent");

    result = await SpecialPowers.spawn(frame.contentWindow, ["#span"], selector => {
      let elem = content.document.querySelector(selector);
      return elem.textContent;
    });

    is(result, "Hello there.", "Got correct element text from window proxy");

    result = await SpecialPowers.spawn(SpecialPowers.getPrivilegedProps(frame, "browsingContext"),
                                       ["#span"], selector => {
      let elem = content.document.querySelector(selector);
      return elem.textContent;
    });

    is(result, "Hello there.", "Got correct element text from browsing context");
  });
</script>
</body>
</html>
